refactor(auth): reorganize auth into cookie/ and open-cloud/ directories
Rename auth/roblox-auth/ to auth/cookie/ and auth/credential-store.ts
to auth/open-cloud/credential-store.ts. Makes the two distinct auth
domains clear: cookie (session auth with CSRF, rotation, platform
stores) vs open-cloud (static API key storage).
