+++ tools/nevermore-cli-helpers/src/auth/cookie/linux.ts
