+++ tools/nevermore-cli-helpers/src/auth/cookie/macos.ts
