+++ tools/nevermore-cli-helpers/src/auth/cookie/validate-cookie.test.ts
