+++ tools/nevermore-cli-helpers/src/auth/cookie/windows.ts
