+++ tools/nevermore-cli-helpers/src/auth/open-cloud/credential-store.ts
